Use a structure to hold the state; refactor code.
Use a structure to hold the state instead of using some globals so we
can have more than one state.

o Refactor code to use the state.
o Add function to initialize a state from a seed.  (This needs a
  better name.)

Test still passes and timing test indicates we're as fast as mt19937
now.